backend/requirements.txt
flask
flask-cors
openai
requests
beautifulsoup4
python-dotenv